requirements.txt
PyQt6
pysnmp-lextudio
zeroconf
requests
